WIP: Added a tabbed view for lights and light groups.
