+++ src/client/components/ModuleWrapper.css
@@ -1,3 +1,6 @@
 .content-wrap {
-    padding: 20px 20px
+    max-width: 480px;
+    margin-left: auto;
+    margin-right: auto;
+    padding: 20px 20px;
 }
