--- src/client/components/lights/LightList.css
@@ -1,1 +1,0 @@
-
