+++ src/client/containers/lights/LightGroupsTab.css
@@ -1,8 +1,5 @@
 .light-group__card {
     margin-bottom: 8px;
-    max-width: 480px;
-    margin-left: auto;
-    margin-right: auto;
 }
 
 .light-group__header {
