+++ src/client/containers/lights/LightsTab.css
@@ -1,0 +1,3 @@
+.lights-tab__card {
+    margin-bottom: 8px;
+}
